+++ src/main/java/com/aviatur/sgia/repository/SedeRepository.java
@@ -6,13 +6,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface SedeRepository extends JpaRepository<Sede, Long> {
+public interface SedeRepository extends JpaRepository<Sede, Integer> {
 
     List<Sede> findAllByOrderByNombreAsc();
 
     boolean existsByNombreIgnoreCase(String nombre);
 
-    boolean existsByNombreIgnoreCaseAndIdNot(String nombre, Long id);
+    boolean existsByNombreIgnoreCaseAndIdNot(String nombre, Integer id);
 
     Optional<Sede> findByNombreIgnoreCase(String nombre);
 }
