+++ src/main/java/com/aviatur/sgia/repository/SistemaOperativoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SistemaOperativoRepository extends JpaRepository<SistemaOperativo, Long> {
+public interface SistemaOperativoRepository extends JpaRepository<SistemaOperativo, Integer> {
 
     boolean existsByNombreIgnoreCaseAndVersionIgnoreCase(String nombre, String version);
