+++ src/main/java/com/aviatur/sgia/repository/SoftwareRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SoftwareRepository extends JpaRepository<Software, Long> {
+public interface SoftwareRepository extends JpaRepository<Software, Integer> {
 
     boolean existsByNombreIgnoreCase(String nombre);
